lots of stuff on this commit
i started with adding in a bunch of region section just for readability and navigatability of the code.

i ofcourse added the glm headers.

i also added a "setMat4()" function to the shader class to set a matrix uniform.

in the current version, two crates are rendered with two separate draw calls from the same VAO and VBO, and they are doing two different things since the transformation matrix is set each frame.