okay, holy crap quite a lot to commit here.
Most of it is just the InputManager class.

the input manager class is a singleton, so it's set up that way and has some pretty off the beaten path semantics involved with it's defintion.

in a nutshell tho, the glfwSetKeyCallback function requires that the function you give it is not a non-static class member function. So the only way I coudl figure out how to have an input system class that has a method for that callback was to make the initial callback method a static method, and build the rest of the InputManager class as a singleton.

long story short it has a static function that will get a reference to the current (only) instance of InputManager and then call the instance method on that.

on the opengl side of things i'm just playing around with the view matrix and stuff. got way too distracted by the input system.

but it's done!!!

https://stackoverflow.com/questions/7676971/pointing-to-a-function-that-is-a-class-member-glfw-setkeycallback

https://stackoverflow.com/questions/1008019/c-singleton-design-pattern