I *believe* I have all of my objects written up correctly and the system should work at this point.
Lets write out a quick run-down:

At the top of the hierarchy is the "World" object. A "World" has a Dir_Light and vectors of Point and Spot Lights, It has a Camera object, and it has a vector of World_Object(s).

It has an instance method called "draw_world" where it will loop through each object in the "world_objects" vector and draw each of them based on it's own fields (position, rotation, etc.), their materials, the lights in the world, and the position of the camera.

World_Object(s) have a material and other essential fields such as position and rotation.

Lights have all the fields you would expect for the type of light; Color, position, direction, etc.