Everything is almsot working!!!
I realized i made a mistake when writing the classes because all of the getters were returning by value and not by reference so any "get_xxxx().set_xxxx()" was not actually changing the internal object, and instead just an instance of it. Remember this.

I honestly just need to go in and refactor all of the classes and just get rid of all the setters and getters if theyre not necessary. Because most of them are just flat out silly.